+++ server/api/socket/initiative.py
@@ -14,7 +14,7 @@ from models.db import db
 from models.role import Role
 from models.shape.access import has_ownership
 from state.game import game_state
-from utils import logger
+from logs import logger
 
 
 class ServerInitiativeEffect(TypedDict):