+++ server/api/socket/label.py
@@ -6,7 +6,7 @@ from api.socket.constants import GAME_NS
 from app import app, sio
 from models import Label, LabelSelection, PlayerRoom, User
 from state.game import game_state
-from utils import logger
+from logs import logger
 
 
 class LabelVisibilityMessage(TypedDict):