+++ CHANGELOG.md
@@ -15,6 +15,7 @@ these changes will usually be stripped from release notes for the public
 -   Door logic can now specify which block settings to toggle
 -   Add double stroke to client viewport
 -   Show campaign loading animation earlier (in dashboard)
+-   [server] Added log rotation
 -   [tech] SyncTo primitive modified to an alternative Sync structure
 -   [tech] Polygon can now have client-side multi-stroke